new traits for master worker handlind
