+++ src/main/scala/com/github/leananeuber/hasher/actors/WorkerManager.scala
@@ -2,9 +2,10 @@ package com.github.leananeuber.hasher.actors
 
 import akka.actor.SupervisorStrategy.Restart
 import akka.actor.{Actor, ActorLogging, ActorRef, Cancellable, OneForOneStrategy, Props, SupervisorStrategy}
-import com.github.leananeuber.hasher.protocols.SessionSetupProtocol.{RegisterAtSession, RegisteredAtSessionAck, SetupSessionConnectionTo}
-import com.github.leananeuber.hasher.actors.password_cracking.PasswordCrackingWorker
+import com.github.leananeuber.hasher.actors.password_cracking.{PasswordCrackingMaster, PasswordCrackingWorker}
 import com.github.leananeuber.hasher.actors.password_cracking.PasswordCrackingWorker.CrackingFailedException
+import com.github.leananeuber.hasher.protocols.MasterWorkerProtocol.SetupConnectionTo
+import com.github.leananeuber.hasher.protocols.SessionSetupProtocol.{RegisterAtSession, RegisteredAtSessionAck, SetupSessionConnectionTo}
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration._
@@ -48,7 +49,10 @@ class WorkerManager(nWorkers: Int) extends Actor with ActorLogging {
       val registerCancellable = context.system.scheduler.schedule(Duration.Zero, 5 seconds) {
         sessionSelection ! RegisterAtSession(nWorkers)
       }
-      context.children.foreach(_ ! SetupSessionConnectionTo(address))
+
+      // for password cracking children (currently there are no other ones)
+      val masterSelection = context.actorSelection(s"$address/user/${Session.sessionName}/${PasswordCrackingMaster.name}")
+      context.children.foreach(_ ! SetupConnectionTo(masterSelection))
       context.become(waitingForSetupAck(registerCancellable))
   }
